requirements.txt
langchain
langchain-google-genai
PyGithub
python-dotenv
streamlit